+++ crates/tabby-scheduler/src/dataset/mod.rs
@@ -68,38 +68,3 @@ pub fn create_dataset(config: &[RepositoryConfig]) {
 
     writer.flush().expect("Failed to flush writer");
 }
-
-pub(crate) mod metrics {
-    use std::cmp::max;
-
-    pub fn max_line_length(content: &str) -> usize {
-        content.lines().map(|x| x.len()).reduce(max).unwrap_or(0)
-    }
-
-    pub fn avg_line_length(content: &str) -> f32 {
-        let mut total = 0;
-        let mut len = 0;
-        for x in content.lines() {
-            len += 1;
-            total += x.len();
-        }
-
-        if len > 0 {
-            total as f32 / len as f32
-        } else {
-            0.0
-        }
-    }
-
-    pub fn alphanum_fraction(content: &str) -> f32 {
-        let num_alphanumn: f32 = content
-            .chars()
-            .map(|x| f32::from(u8::from(x.is_alphanumeric())))
-            .sum();
-        if !content.is_empty() {
-            num_alphanumn / content.len() as f32
-        } else {
-            0.0
-        }
-    }
-}
